+++ cli/cdkscale.c
@@ -1,4 +1,4 @@
-/* $Id: cdkscale.c,v 1.10 2008/11/04 23:36:39 tom Exp $ */
+/* $Id: cdkscale.c,v 1.11 2011/05/16 22:58:54 tom Exp $ */
 
 #include <cdk_test.h>
 
@@ -169,7 +169,7 @@ int main (int argc, char **argv)
    {
       /* Split the button list up. */
       buttonList = CDKsplitString (buttons, '\n');
-      buttonCount = CDKcountStrings (buttonList);
+      buttonCount = (int)CDKcountStrings (buttonList);
 
       /* We need to create a buttonbox widget. */
       buttonWidget = newCDKButtonbox (cdkScreen,
