+++ cli/cdkscroll.c
@@ -1,4 +1,4 @@
-/* $Id: cdkscroll.c,v 1.9 2005/12/27 17:29:58 tom Exp $ */
+/* $Id: cdkscroll.c,v 1.10 2011/05/16 22:58:54 tom Exp $ */
 
 #include <cdk_test.h>
 
@@ -111,7 +111,7 @@ int main (int argc, char **argv)
    {
       /* Split the scroll lines up. */
       scrollList = CDKsplitString (list, '\n');
-      scrollLines = CDKcountStrings (scrollList);
+      scrollLines = (int)CDKcountStrings (scrollList);
    }
 
    /* Start curses. */
@@ -163,7 +163,7 @@ int main (int argc, char **argv)
    {
       /* Split the button list up. */
       buttonList = CDKsplitString (buttons, '\n');
-      buttonCount = CDKcountStrings (buttonList);
+      buttonCount = (int)CDKcountStrings (buttonList);
 
       /* We need to create a buttonbox widget. */
       buttonWidget = newCDKButtonbox (cdkScreen,
