+++ cli/cdkslider.c
@@ -1,4 +1,4 @@
-/* $Id: cdkslider.c,v 1.10 2008/11/04 23:36:39 tom Exp $ */
+/* $Id: cdkslider.c,v 1.11 2011/05/16 22:59:03 tom Exp $ */
 
 #include <cdk_test.h>
 
@@ -180,7 +180,7 @@ int main (int argc, char **argv)
    {
       /* Split the button list up. */
       buttonList = CDKsplitString (buttons, '\n');
-      buttonCount = CDKcountStrings (buttonList);
+      buttonCount = (int)CDKcountStrings (buttonList);
 
       /* We need to create a buttonbox widget. */
       buttonWidget = newCDKButtonbox (cdkScreen,
